==> pgms -GIT/00.Libname.R <==
#---------------------------------------------------------------------------------------------------------------------------------------
#  _____  _                     _    ______  _         _      _                _   _           ______             _              _   
# /  ___|| |                   | |   |  ___|(_)       | |    | |              | | | |          | ___ \           (_)            | |  
# \ `--. | |_  _ __  ___   ___ | |_  | |_    _   __ _ | |__  | |_  ___  _ __  | | | |  ______  | |_/ /_ __  ___   _   ___   ___ | |_ 
#  `--. \| __|| '__|/ _ \ / _ \| __| |  _|  | | / _` || '_ \ | __|/ _ \| '__| | | | | |______| |  __/| '__|/ _ \ | | / _ \ / __|| __|
# /\__/ /| |_ | |  |  __/|  __/| |_  | |    | || (_| || | | || |_|  __/| |    \ \_/ /          | |   | |  | (_) || ||  __/| (__ | |_ 
# \____/  \__||_|   \___| \___| \__| \_|    |_| \__, ||_| |_| \__|\___||_|     \___/           \_|   |_|   \___/ | | \___| \___| \__|
#                                               __/ |                                                          _/ |                 
#                                              |___/                                                          |__/                  
#---------------------------------------------------------------------------------------------------------------------------------------

# --------------------------------------------------------------------------------
# --- Libnames Utilizadas ---------------------------------------
# --------------------------------------------------------------------------------

# Libs
#source("E:/Projetos R/Street Figther/pgms -GIT/00.Libname.R")

library(XML)
library(devtools)
library(stringr)
library(imager)
library(dplyr)
library(tesseract)
library(magick)
library(rvest)

library(ggplot2)
library(reshape2)
library(png)
library(grid)
require(gridExtra)
library(wesanderson)
library(viridis)          
library(reshape)   
library(reshape2)   

library(ggcorrplot)
library(RColorBrewer)
